# perf(web): Optimize `TextDecoder` by adding a new `U16String` type