--- a/serde_v8/src/lib.rs
+++ b/serde_v8/src/lib.rs
@@ -14,6 +14,7 @@ pub use keys::KeyCache;
 pub use magic::buffer::MagicBuffer as Buffer;
 pub use magic::bytestring::ByteString;
 pub use magic::string_or_buffer::StringOrBuffer;
+pub use magic::u16string::U16String;
 pub use magic::Value;
 pub use ser::{to_v8, Serializer};
 pub use serializable::{Serializable, SerializablePkg};