Before:
```
Benchmark #10: deno run cli/tests/testdata/text_decoder_perf.js 
  Time (mean ± σ):      90.1 ms ±   1.3 ms    [User: 90.1 ms, System: 7.9 ms]
  Range (min … max):    87.7 ms …  94.1 ms    32 runs
```

After:
```
Benchmark #10: deno run cli/tests/testdata/text_decoder_perf.js 
  Time (mean ± σ):      51.5 ms ±   0.7 ms    [User: 53.4 ms, System: 8.2 ms]
  Range (min … max):    50.0 ms …  53.0 ms    57 runs
```

﻿Closes #10978.

<!--
Before submitting a PR, please read http://deno.land/manual/contributing

1. Give the PR a descriptive title.

  Examples of good title:
    - fix(std/http): Fix race condition in server
    - docs(console): Update docstrings
    - feat(doc): Handle nested reexports

  Examples of bad title:
    - fix #7123
    - update docs
    - fix bugs

2. Ensure there is a related issue and it is referenced in the PR text.
3. Ensure there are tests that cover the changes.
4. Ensure `cargo test` passes.
5. Ensure `./tools/format.js` passes without changing files.
6. Ensure `./tools/lint.js` passes.
-->